gtk: split into 3.x and 4.x recipes
