+++ recipes/gtk/3.x/conandata.yml
@@ -1,0 +1,4 @@
+sources:
+  "3.24.43":
+    url: "https://download.gnome.org/sources/gtk+/3.24/gtk+-3.24.43.tar.xz"
+    sha256: "7e04f0648515034b806b74ae5d774d87cffb1a2a96c468cb5be476d51bf2f3c7"
