+++ recipes/gtk/4.x/conandata.yml
@@ -5,6 +5,3 @@ sources:
   "4.7.0":
     url: "https://download.gnome.org/sources/gtk/4.7/gtk-4.7.0.tar.xz"
     sha256: "913fcd9d065efb348723e18c3b9113e23b92072e927ebd2f61d32745c8228b94"
-  "3.24.43":
-    url: "https://download.gnome.org/sources/gtk+/3.24/gtk+-3.24.43.tar.xz"
-    sha256: "7e04f0648515034b806b74ae5d774d87cffb1a2a96c468cb5be476d51bf2f3c7"
